making changes to accept float values while buying/selling coins (#118)
When we input a float value suppose say 0.01,the parseInt value is set
to true(because it cannot handle float values,it truncates the digits
after decimal points and returns 0) and the button is disabled even when
the transaction is valid.
#117

+++ src/routes/stablecoin.jsx
@@ -267,8 +267,8 @@ export default function Stablecoin() {
     : sellValidity === TRANSACTION_VALIDITY.OK;
 
   const buttonDisabled =
-    isNaN(parseInt(value)) ||
-    parseInt(value) === 0 ||
+    isNaN(parseFloat(value)) ||
+    parseFloat(value) === 0 ||
     isWrongChain ||
     !transactionValidated;
